Rename speak to synthesize
