+++ packages/component/src/createSynthesize.spec.js
